Renamed and moved some stuff around to improve readability.
