+++ src/Inkluzitron/Data/Entities/QuizDoubleItem.cs
@@ -1,4 +1,4 @@
-﻿namespace Inkluzitron.Data
+﻿namespace Inkluzitron.Data.Entities
 {
     public class QuizDoubleItem : QuizItem
     {
