+++ src/Inkluzitron/Extensions/EmojiParsingExtensions.cs
@@ -1,9 +1,8 @@
 ﻿using Discord;
-using System.Globalization;
 
 namespace Inkluzitron.Extensions
 {
-    public static class EmojiParsingExtensions
+    static public class EmojiParsingExtensions
     {
         static public IEmote ToDiscordEmote(this string configuredValue)
         {
